@@ -37,7 +37,7 @@
     "predist": "echo hybrid",
     "dist": "npm run dist:web && npm run dist:python",
     "compile": "tsc",
-    "lint": "tslint -c tslint.json -p .",
+    "lint": "tslint -c tslint.json -p . 'src/**/*.ts?(x)' 'tests/**/*.ts?(x)'",
     "docs": "npm run docs:web && npm run docs:python",
     "prebuild": "echo hybrid",
     "posttest": "echo hybrid",
@@ -68,20 +68,21 @@
     "prebuild:python": "node -e \"process.exit(process.env.PHOVEA_SKIP_TESTS === undefined?1:0)\" || npm run test:python",
     "prebuild:web": "node -e \"process.exit(process.env.PHOVEA_SKIP_TESTS === undefined?1:0)\" || npm run test:web"
   },
-  "dependencies": {
-    "tdp_gene": "github:Caleydo/tdp_gene#develop"
-  },
-  "optionalDependencies": {
-    "ordino": "github:Caleydo/ordino#develop",
-    "dTiles": "github:datavisyn/dTiles#develop"
+  "main": "build/tdp_publicdb.js",
+  "engines": {
+    "npm": ">= 6.12",
+    "node": ">= 12.13"
   },
   "devDependencies": {
     "@types/jasmine": "2.5.47",
     "awesome-typescript-loader": "3.1.2",
+    "cache-loader": "1.2.0",
     "css-loader": "0.28.0",
-    "extract-loader": "0.1.0",
     "extract-text-webpack-plugin": "2.1.0",
+    "extract-loader": "0.1.0",
+    "ifdef-loader": "2.0.0",
     "file-loader": "0.11.1",
+    "fork-ts-checker-webpack-plugin": "0.4.1",
     "html-loader": "0.4.5",
     "imports-loader": "0.7.1",
     "jasmine": "2.5.3",
@@ -99,17 +100,21 @@
     "raw-loader": "0.5.1",
     "sass-loader": "6.0.7",
     "style-loader": "0.16.1",
+    "thread-loader": "1.1.2",
+    "ts-loader": "4.0.1",
     "tslib": "1.9.0",
     "tslint": "5.9.1",
     "typedoc": "0.11.1",
-    "typescript": "2.7.2",
+    "typescript": "2.8.1",
     "url-loader": "0.5.8",
     "webpack": "2.3.3",
-    "webpack-dev-server": "2.4.2",
-    "cache-loader": "1.2.0",
-    "ifdef-loader": "2.0.0",
-    "fork-ts-checker-webpack-plugin": "0.4.1",
-    "thread-loader": "1.1.2",
-    "ts-loader": "4.0.1"
+    "webpack-dev-server": "2.4.2"
+  },
+  "dependencies": {
+    "tdp_gene": "github:Caleydo/tdp_gene#develop"
+  },
+  "optionalDependencies": {
+    "ordino": "github:Caleydo/ordino#develop",
+    "dTiles": "github:datavisyn/dTiles#develop"
   }
 }
