@@ -3,8 +3,6 @@
  */
 
 const spawnSync = require('child_process').spawnSync;
-const path = require('path');
-const resolve = path.resolve;
 const fs = require('fs');
 
 function gitHead(cwd) {
@@ -26,7 +24,7 @@ function resolvePlugin(repo, version) {
     }
   }
   // not a git repo
-  return version
+  return version;
 }
 
 function toVersion(v) {
@@ -35,8 +33,8 @@ function toVersion(v) {
   const fmt = now
     .replace(/T/, ' ')
     .replace(/\..+/, '')
-    .replace(/[-:]/,'')
-    .replace(' ','-');
+    .replace(/[-:]/, '')
+    .replace(' ', '-');
   return v.replace('SNAPSHOT', fmt);
 }
 
@@ -55,7 +53,7 @@ function _main() {
     repository: (pkg.repository || {}).url
   };
 
-  const l = ('build/source/' + name).split('/');
+  const l = ('build/source/' + name.toLowerCase()).split('/');
   l.forEach((_, i) => {
     const path = l.slice(0, i + 1).join('/');
     if (!fs.existsSync(path)) {
@@ -63,9 +61,8 @@ function _main() {
     }
   });
 
-  fs.writeFileSync('build/source/' + name + '/buildInfo.json', JSON.stringify(buildInfo, null, ' '));
+  fs.writeFileSync('build/source/' + name.toLowerCase() + '/buildInfo.json', JSON.stringify(buildInfo, null, ' '));
 }
-
 
 if (require.main === module) {
   _main();
